requirements.txt
flask
cassandra-driver
python-dotenv
flask-cors